refacto: split to lib + layout
